+++ pyiron_base/storage/helper_functions.py
@@ -1,0 +1,42 @@
+import h5io
+from pyiron_base.utils.error import retry
+
+
+def read_hdf5(fname, title="h5io", slash="ignore"):
+    return retry(
+        lambda: h5io.read_hdf5(
+            fname=fname,
+            title=title,
+            slash=slash,
+        ),
+        error=BlockingIOError,
+        msg=f"Two or more processes tried to access the file {fname}.",
+        at_most=10,
+        delay=1,
+    )
+
+
+def write_hdf5(
+    fname,
+    data,
+    overwrite=False,
+    compression=4,
+    title="h5io",
+    slash="error",
+    use_json=False,
+):
+    retry(
+        lambda: h5io.write_hdf5(
+            fname=fname,
+            data=data,
+            overwrite=overwrite,
+            compression=compression,
+            title=title,
+            slash=slash,
+            use_json=use_json,
+        ),
+        error=BlockingIOError,
+        msg=f"Two or more processes tried to access the file {fname}.",
+        at_most=10,
+        delay=1,
+    )
